Split grid generation and rendering
