+++ src/lib.rs
@@ -6,6 +6,8 @@ pub mod pngs {
 
 pub mod constants;
 pub mod generator;
+pub mod grid_generator;
+pub mod grid_renderer;
 pub mod service;
 pub mod sheets;
 pub mod spelunkicon;
